+++ templates/TESTPAGE.html
@@ -6,18 +6,12 @@
     <script>
       $(() => {
         console.log("READY");
-        // $.ajax({
-        //   type: "GET",
-        //   url: "/movie",
-        //   data: {},
-        //   success: (response) => {
-        //     console.log(response["result"]);
-        //   },
-        // });
+
       });
     </script>
   </head>
   <body>
     <h1>TESTTEST</h1>
+    {{ a["name"] }}
   </body>
 </html>
