+++ user.py
@@ -1,0 +1,82 @@
+from flask import Blueprint, render_template, request, jsonify, make_response, redirect, url_for
+from pymongo import MongoClient
+
+import jwt
+import hashlib
+import os
+from datetime import datetime, timedelta
+from dotenv import load_dotenv
+
+from regex import *
+
+
+load_dotenv()
+URL = os.environ.get("MongoDB_URL")
+KEY = os.environ.get("SECRET_KEY")
+
+client = MongoClient(URL, tls=True, tlsAllowInvalidCertificates=True)
+db = client.spamovie
+
+user_bp = Blueprint("user", __name__)
+
+
+
+@user_bp.route("/signin")
+def components_sign_in():
+   return render_template("components/sign_in.html")
+
+
+@user_bp.route("/signin", methods=["POST"])
+def sign_in():
+   username = request.form["username"]
+   password = request.form["password"]
+   if name_check(username) is not True:
+      return jsonify({"msg": "아이디 형식은 알파벳,한글,숫자 3~15자 입니다."})
+   if pass_check(password) is not True:
+      return jsonify({"msg": "비밀번호 형식은 알파벳,숫자 8~15자 입니다."})
+
+   password_hash = hashlib.sha256(password.encode("utf-8")).hexdigest()
+   user = db.users.find_one({"username": username, "password": password_hash})
+   if user is not None:
+      uid = user["uid"]
+      payload = {
+         "uid": uid,
+         "username": username,
+         "exp": datetime.utcnow() + timedelta(seconds = 60*60)
+      }
+      token = jwt.encode(payload, KEY, algorithm="HS256") #.decode("utf-8")   # annotate while running in localhost
+      cookie = make_response()
+      cookie.set_cookie("logintoken", token)
+      return cookie
+   else:
+      return jsonify({"msg": "아이디, 비밀번호가 틀렸습니다."})
+
+
+@user_bp.route("/signup")
+def components_sign_up():
+   return render_template("components/sign_up.html")
+
+
+@user_bp.route("/signup", methods=["POST"])
+def sign_up():
+   username = request.form["username"]
+   password = request.form["password"]
+   if name_check(username) is not True:
+      return jsonify({"msg": "아이디 형식은 알파벳,한글,숫자 3~15자 입니다."})
+   if pass_check(password) is not True:
+      return jsonify({"msg": "비밀번호 형식은 알파벳,숫자 8~15자 입니다."})
+
+   cnt = db.users.find_one({}, {"_id": False})
+   uid = cnt["cnt"] + 1
+   db.users.update_one({}, {"$set": {"cnt": uid}})
+
+   password_hash = hashlib.sha256(password.encode("utf-8")).hexdigest()
+   profile = {
+      "uid": uid,
+      "username": username,
+      "password": password_hash,
+      "reviews": []
+   }
+   db.users.insert_one(profile)
+
+   return redirect(url_for("/sign_in"))
